+++ BooleanExpressionEvaluationTests/Expression/ExpressionTests.swift
@@ -26,17 +26,17 @@ class ExpressionTests: XCTestCase {
     // MARK: - Functions
 
     func testInit_Minimum() {
-        let expression: [ExpressionElement] = [.operand(.variable("variable")), .comparisonOperator(.equal), .operand(.number(2))]
+        let expression: Expression = [.operand(.variable("variable")), .comparisonOperator(.equal), .operand(.number(2))]
         test("variable == 2", expression)
     }
 
     func testInit_MinimumWithExtraSpace() {
-        let expression: [ExpressionElement] = [.operand(.variable("variable")), .comparisonOperator(.equal), .operand(.number(2))]
+        let expression: Expression = [.operand(.variable("variable")), .comparisonOperator(.equal), .operand(.number(2))]
         test("variable  == 2", expression)
     }
 
     func testInit_MinimumWithExtraSpaceAfterBracket() {
-        let expression: [ExpressionElement] = [.bracket(.opening), .operand(.variable("variable")), .comparisonOperator(.equal), .operand(.number(2)), .bracket(.closing)]
+        let expression: Expression = [.bracket(.opening), .operand(.variable("variable")), .comparisonOperator(.equal), .operand(.number(2)), .bracket(.closing)]
         test("( variable == 2)", expression)
     }
 
@@ -64,8 +64,8 @@ class ExpressionTests: XCTestCase {
             _ = try Expression(string)
             XCTFail("Initializing the expression with an incorrect string should throw an error")
         } catch {
-            guard case ExpressionError.incorrectElement(_) = error else {
-                XCTFail("Initizalieing an Expression with an empty string should have thrown a ExpressionError.emptyExpression error")
+            guard case ExpressionError.invalidVariableName(_) = error else {
+                XCTFail("Initizalizing an Expression with an empty string should have thrown a ExpressionError.invalidVariableName error")
                 return
             }
         }
@@ -81,6 +81,56 @@ class ExpressionTests: XCTestCase {
                 return
             }
         }
+    }
+
+    func testInit_StringWithSpace() {
+        let string = #"variable == "String with space""#
+        do {
+            let expression = try Expression(string)
+            XCTAssertEqual(expression, [.operand(.variable("variable")), .comparisonOperator(.equal), .operand(.string("String with space"))])
+        } catch {
+            XCTFail(error.localizedDescription)
+        }
+    }
+
+    func testInit_StringExtraSpaces() {
+        let string = #"variable == " String with space " "#
+        do {
+            let expression = try Expression(string)
+            XCTAssertEqual(expression, [.operand(.variable("variable")), .comparisonOperator(.equal), .operand(.string(" String with space "))])
+        } catch {
+            XCTFail(error.localizedDescription)
+        }
+    }
+
+    // MARK: Codable
+
+    func testCanDecodeFromString() {
+        guard let data = try? JSONEncoder().encode("(variable >= 1.5)") else {
+            XCTFail("Unable to encode a String ? Strange 🤔")
+            return
+        }
+        let decoder = JSONDecoder()
+        let expression = try? decoder.decode(Expression.self, from: data)
+
+        XCTAssertEqual(expression, [.bracket(.opening), .operand(.variable("variable")), .comparisonOperator(.greaterThanOrEqual), .operand(.number(1.5)), .bracket(.closing)])
+    }
+
+    func testCanEncodeToString() {
+
+        let expression: Expression = [.bracket(.opening), .operand(.variable("variable")), .comparisonOperator(.greaterThanOrEqual), .operand(.number(1.5)), .bracket(.closing)]
+        let data: Data
+        do {
+            data = try JSONEncoder().encode(expression)
+        } catch {
+            XCTFail("Unable to encode the expression \(expression.description): \(error.localizedDescription)")
+            return
+        }
+
+        let decoder = JSONDecoder()
+        let stringExpression = try? decoder.decode(String.self, from: data)
+
+        XCTAssertEqual(stringExpression, expression.description)
     }
 
     // MARK: Helpers